Refactor into modular monitor package
