+++ monitor/__init__.py
@@ -1,0 +1,8 @@
+"""Monitoring utilities package."""
+
+__all__ = [
+    "snmp",
+    "network",
+    "web",
+    "ups",
+]
